+++ libraries/customisations/src/main/kotlin/com/bumble/appyx/utils/customisations/MutableNodeCustomisationDirectory.kt
@@ -4,7 +4,15 @@ import kotlin.reflect.KClass
 
 interface MutableNodeCustomisationDirectory : NodeCustomisationDirectory {
 
+    @Deprecated(
+        message = "Use putSubDirectoryLazy to avoid potential performance issues",
+        replaceWith = ReplaceWith("putSubDirectoryLazy(key) { value }")
+    )
     fun <T : Any> putSubDirectory(key: KClass<T>, value: NodeCustomisationDirectory)
 
+    @Deprecated(
+        message = "Use putLazy to avoid potential performance issues",
+        replaceWith = ReplaceWith("putLazy(key) { value }")
+    )
     fun <T : NodeCustomisation> put(key: KClass<T>, value: T)
 }